pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "op-orm"
version = "0.1.0"
dependencies = [
    "onepassword-sdk",
    "jinja2",
]
requires-python = ">=3.12"
authors = [
    {name = "Ilke", email = "github@ilke.dev"},
]
maintainers = [
    {name = "Ilke", email = "github@ilke.dev"},
]
description = "Object Relational Mapping for 1Password items, manage 1Password vaults with code."
readme = "README.md"
keywords = ["1password", "orm", "op", "op-orm", "onepassword"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3.12",
]

[tool.setuptools]
packages = ["op_orm"]  # Declare the actual package name
package-dir = {"op_orm" = "op_orm"}  # Declare the source directory

[project.optional-dependencies]
dev = [
    "pytest",
    "debugpy",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpaths = [
    "op_orm",
]
testpaths = [
    "tests",
]

[project.scripts]
generate_deployments = "op_orm.generate_deployments:main"